runfix: move font size state to its own file
