+++ src/script/page/MainContent/panels/preferences/OptionPreferences.tsx
@@ -26,13 +26,13 @@ import {container} from 'tsyringe';
 import {Checkbox, CheckboxLabel, RangeInput} from '@wireapp/react-ui-kit';
 import {WebAppEvents} from '@wireapp/webapp-events';
 
-import {RootFontSize, useAppMainState} from 'src/script/page/state';
 import {useKoSubscribableChildren} from 'Util/ComponentUtil';
 import {t} from 'Util/LocalizerUtil';
 
 import {PreferencesPage} from './components/PreferencesPage';
 import {PreferencesRadio} from './components/PreferencesRadio';
 import {PreferencesSection} from './components/PreferencesSection';
+import {RootFontSize, useFontSizeState} from './useFontSizePreferenceState';
 
 import {PropertiesRepository} from '../../../../properties/PropertiesRepository';
 import {PROPERTIES_TYPE} from '../../../../properties/PropertiesType';
@@ -68,6 +68,7 @@ const OptionPreferences: React.FC<OptionPreferencesProps> = ({
   const [optionDarkMode, setOptionDarkMode] = useState<boolean>(settings.interface.theme === ThemeViewModelThemes.DARK);
   const [optionSendPreviews, setOptionSendPreviews] = useState<boolean>(settings.previews.send);
   const [optionNotifications, setOptionNotifications] = useState<NotificationPreference>(settings.notifications);
+  const {currentRootFontSize, setCurrentRootFontSize} = useFontSizeState(state => state.rootFontSize);
 
   useEffect(() => {
     const updateProperties = ({settings}: WebappProperties): void => {
@@ -88,8 +89,6 @@ const OptionPreferences: React.FC<OptionPreferencesProps> = ({
       amplify.unsubscribe(WebAppEvents.PROPERTIES.UPDATED, updateProperties);
     };
   }, []);
-
-  const {currentRootFontSize, setCurrentRootFontSize} = useAppMainState(state => state.rootFontSize);
 
   const sliderValue = fontSizes.indexOf(currentRootFontSize);
